ml/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
claude-code-sdk==0.0.25
click==8.3.1
cryptography==46.0.3
fastapi==0.121.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.22.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
referencing==0.37.0
rpds-py==0.29.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
